requirements.txt
# Core Discord Bot
discord.py>=2.3.2

# Audio/Music
yt-dlp>=2024.1.0
PyNaCl>=1.5.0
davey>=0.1.0

# HTTP Requests
aiohttp>=3.9.0
requests>=2.31.0

# Spotify Integration
spotipy>=2.22.1

# Web Scraping (Lyrics)
beautifulsoup4>=4.12.0

# Configuration
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
